si inchioda con più di 10 particelle

+++ src/BarnesHut.h
@@ -9,18 +9,18 @@
 #define BarnesHut_h_
 
 int nbodies = 1000;
-double dt = 0.01;
+long double dt = 0.01;
 int steps = 25;
-double eps = 0.0;
-double tol = 0.5;
+long double eps = 0.0;
+long double tol = 0.5;
 
-double dthf = 0.0;
-double epssq = 0.0;
-double itolsq = 0.0;
+long double dthf = 0.0;
+long double epssq = 0.0;
+long double itolsq = 0.0;
 
 typedef struct leaf_t{
-	double vel[3];
-	double acc[3];
+	long double vel[3];
+	long double acc[3];
 } leaf_t;
 
 //typedef struct internal_node_t{
@@ -42,8 +42,8 @@ typedef union cell{
 // type definisce il tipo espresso dalla union: 0 = body_t, 1 = internal_node_t
 typedef struct node_t{
 	int type;
-	double mass;
-	double pos[3];
+	long double mass;
+	long double pos[3];
 	cell cell;
 } node_t;
